+++ src/PubSub.Domain/EventBus/Models/Stock.cs
@@ -1,6 +1,6 @@
 namespace PubSub.Domain.EventBus.Models;
 
-public abstract record Stock(int Price = 0, string Symbol = null, string Name = null)
+public abstract record Stock(int Price = 0, string Symbol = "", string Name = "")
 {
     public int Price { get; set; } = Price;
     public string Symbol { get; set; } = Symbol;