+++ src/PubSub.Domain/EventBus/Subscribers/GoogleSubscriber.cs
@@ -11,7 +11,7 @@ public class GoogleSubscriber : Subscriber<Google>, IGoogleSubscriber
     {
     }
 
-    public override void Notify(Google message)
+    public override void Handle(Google message)
     {
         Console.WriteLine($"{message.Symbol,-5}-->{message.Name,7} new price is {message.Price}");
     }